greeting_mcp_server/pyproject.toml
[project]
name = "greeting-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=43.0.0",
    "fastapi>=0.115.0",
    "fastmcp>=2.14.1",
    "mcp[cli]>=1.24.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.1",
    "pyjwt[crypto]>=2.10.0",
    "uvicorn[standard]>=0.32.0",
]
